refactor: use-infinite-scroll-observer로 파일 이름 변경 (#447)
