+++ src/hooks/use-infinite-scroll-observer.ts
